refactor: Cdc, move the remaining code in the root entry point
