+++ integration-libs/cdc/cdc.module.ts
@@ -1,0 +1,8 @@
+import { NgModule } from '@angular/core';
+import { CdcComponentsModule } from '@spartacus/cdc/components';
+import { CdcCoreModule } from '@spartacus/cdc/core';
+
+@NgModule({
+  imports: [CdcComponentsModule, CdcCoreModule],
+})
+export class CdcModule {}
