+++ integration-libs/cdc/components/gigya-raas/gigya-raas.component.spec.ts
@@ -1,8 +1,7 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
-import { CdcJsService } from '@spartacus/cdc';
 import { GigyaRaasComponentData } from '@spartacus/cdc/core';
-import { CdcConfig } from '@spartacus/cdc/root';
+import { CdcConfig, CdcJsService } from '@spartacus/cdc/root';
 import {
   BaseSiteService,
   CmsComponent,
