+++ integration-libs/cdc/components/gigya-raas/gigya-raas.component.ts
@@ -5,9 +5,8 @@ import {
   OnInit,
   ViewEncapsulation
 } from '@angular/core';
-import { CdcJsService } from '@spartacus/cdc';
 import { GigyaRaasComponentData } from '@spartacus/cdc/core';
-import { CdcConfig } from '@spartacus/cdc/root';
+import { CdcConfig, CdcJsService } from '@spartacus/cdc/root';
 import { BaseSiteService, LanguageService, WindowRef } from '@spartacus/core';
 import { CmsComponentData } from '@spartacus/storefront';
 import { Observable } from 'rxjs';
