+++ integration-libs/cdc/ng-package.json
@@ -5,11 +5,7 @@
     "entryFile": "./public_api.ts",
     "umdModuleIds": {
       "@spartacus/core": "core",
-      "@spartacus/storefront": "storefront",
-      "@spartacus/asm/root": "asm/root",
-      "rxjs": "rxjs",
-      "@ngrx/store": "store",
-      "@ngrx/effects": "effects"
+      "@spartacus/storefront": "storefront"
     }
   }
 }
