+++ integration-libs/cdc/public_api.ts
@@ -1,7 +1,1 @@
-/*
- * Public API Surface for CDC
- */
-export * from './src/auth/facade/index';
-export * from './src/auth/store/actions/index';
-export * from './src/cdc.module';
-
+export * from './cdc.module';
