+++ integration-libs/cdc/root/auth/facade/cdc-auth.service.ts
