+++ integration-libs/cdc/root/auth/facade/index.ts
