+++ integration-libs/cdc/root/index.ts
@@ -1,2 +1,4 @@
+export * from './auth/facade/index';
+export * from './auth/store/actions/index';
 export * from './cdc-root.module';
 export * from './config';
